LayoutTests/platform/ios-simulator/fast/media/mq-any-hover-styling-expected.txt

Test the any-hover media feature to guard style rules.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


getComputedStyle(target)["color"] = rgb(255, 0, 0)
getComputedStyle(target)["background-color"] = rgb(0, 0, 0)
getComputedStyle(target)["width"] = 100px
getComputedStyle(target)["height"] = 10px
PASS successfullyParsed is true

TEST COMPLETE
